feat: Improve filtering out of unpublished datasets
Filter out unpublished datasets in the after_parsing method of StadtzhLosdHarvester. This is called in the gather_stage, before we start gathering datasets and creating harvest objects for them.

This means we don't have to change the status of the harvest objects for unpublished datasets to 'delete' later on in the import stage. This way, the output of the job will not misleadingly say we deleted datasets when in fact, the datasets were never harvested.