fix: patch read_datasets_from_db
When a dataset is harvested (imported or changed) via the LOSD-Harvester it works as expected.
As soon as a dataset is changed outside the harvesting context (e.g. CKAN GUI) the guid disappears from the package_extras.
This way we have a fallback, in case the default check for an existing dataset fails.